requirements.txt
pandas
numpy
matplotlib
seaborn
yfinance
numpy-financial
vectorbt
streamlit
google-genai
notebook
ipykernel
